refactored socket connection from main to Shinkansen
