+++ Makefile
@@ -1,9 +1,13 @@
-server: server.o utils.o 
-		g++ -o build/server build/server.o build/Utils.o
+server: server.o Utils.o Shinkansen.o
+		g++ -o build/server build/server.o build/Utils.o build/Shinkansen.o
 
 server.o: 
 		g++ src/server.cpp -c -o build/server.o
-utils.o:
+
+Shinkansen.o:
+	g++ src/Shinkansen.cpp -c -o build/Shinkansen.o
+
+Utils.o:
 		g++ src/Utils.cpp -c -o build/Utils.o
