+++ src/Shinkansen.h
@@ -1,21 +1,10 @@
-#include <stdio.h>
-#include <sys/socket.h>
-#include <unistd.h>
-#include <stdlib.h>
-#include <netinet/in.h>
-#include <string.h>
-#include <iostream>
-#include "Utils.h"
-
-using namespace std;
-
-#define PORT 8080
+#pragma once
 
 class Shinkansen
 {
 private:
-  /* data */
+  int port; 
+  void Initialize();
 public:
-  Shinkansen(/* args */);
-  ~Shinkansen();
+  Shinkansen(int port);
 };
